+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Added
+- `validateWithoutSnakeCased` flag for validators `validateCreatorDetails` method
+
 ## [1.8.0-rc.1] - 2019-06-17
 #### Added
 - `isEmptyBodyAllowed` flag for api-caller `_call` method